rewrote sidebar to angular. added support for custom angular sidebar components. sidebar is now flexbox rather than absolute positoned, and employes position:sticky instead of javascript onscroll hack. a rudimentary eventbus has been put in place, but must be replaced with something more robust.
