+++ app/scripts/app.js
@@ -2,6 +2,7 @@
 
 import jStorage from "../lib/jstorage"
 import { kwicPagerName, kwicPager } from "./components/pager"
+import { sidebarName, sidebarComponent } from "./components/sidebar"
 import { setDefaultConfigValues } from "./settings.js"
 
 setDefaultConfigValues()
@@ -40,7 +41,7 @@ window.korpApp = angular.module("korpApp", [
     "angular.filter",
 ])
 
-korpApp.component(kwicPagerName, kwicPager)
+korpApp.component(kwicPagerName, kwicPager).component(sidebarName, sidebarComponent)
 
 korpApp.config((tmhDynamicLocaleProvider) =>
     tmhDynamicLocaleProvider.localeLocationPattern("translations/angular-locale_{{locale}}.js")
