+++ app/scripts/components/sidebar.scss
@@ -1,0 +1,66 @@
+/** @format */
+.sidebar {
+    width: 250px;
+    &:empty {
+        padding: 0;
+        border: none;
+    }
+    div {
+        margin-bottom: 10px;
+    }
+    p {
+        margin: 0 0 0 0;
+    }
+    ul {
+        margin-top: 0;
+        margin-bottom: 2px;
+        padding-left: 20px;
+        list-style: none;
+        line-height: 16px;
+    }
+
+    #selected_sentence > h4:last-child {
+        display: none;
+    }
+
+    #selected_word > h4:last-child {
+        display: none;
+    }
+    .top-10 {
+        top: 10px;
+    }
+}
+
+.external_link {
+    display: none;
+}
+
+.sidebar li:hover .external_link {
+    background-image: url("../../img/fish_icon.png");
+    background-origin: content-box;
+    background-repeat: no-repeat;
+    display: inline-block !important;
+    vertical-align: middle;
+    height: 12px;
+    width: 20px;
+    background-size: 20px;
+    margin-left: 2em;
+    position: relative;
+    cursor: pointer;
+}
+
+.sidebar_info {
+    display: inline-block;
+    vertical-align: middle;
+    position: relative;
+    top: 1px;
+    left: 5px;
+    opacity: 0.8;
+}
+
+.kwic_sidebar {
+    background-color: #fff;
+    border-style: solid;
+    border-width: 1px;
+    border-color: #999;
+}
