+++ app/styles/styles.scss
@@ -1119,101 +1119,17 @@ table {
 }
 
 #columns {
-  position: relative;
-  max-height: 600px;
+  transition : width 300ms;
+  width: 100%;
+  will-change: width;
+  &:not(.sidebar_visible) {
+    width: calc(100% + 258px);
+    // width: 130%;
+  }
 }
 
 #left-column {
-  position: absolute;
-  left: 0px;
-  right : 0;
-  transition : right 300ms;
-  &.sidebar_visible {
-    right : 258px;
-    transition : right 300ms;
-  }
-}
-
-#sidebar {
-  position: absolute;
-  right: 0px;
-  width: 250px;
-  padding: 5px 0px 5px 5px;
-  right : -268px;
-  transition : right 300ms;
-  &.visible {
-    transition : right 300ms;
-    right : 0;
-  }
-  &.visible.fixed {
-    right: 8px;
-    position: fixed;
-    top: 8px;
-    overflow-y: auto;
-    max-height: calc(100vh - 12px);
-  }
-  &:empty {
-    padding: 0;
-    border: none;
-  }
-}
-
-.external_link {
-  display: none;
-}
-
-#sidebar li:hover .external_link {
-  background-image: url("../img/fish_icon.png");
-  background-origin: content-box;
-  background-repeat: no-repeat;
-  display: inline-block !important;
-  vertical-align: middle;
-  height: 12px;
-  width: 20px;
-  background-size: 20px;
-  margin-left: 2em;
-  position: relative;
-  cursor: pointer;
-}
-
-.sidebar_info {
-  display: inline-block;
-  vertical-align: middle;
-  position: relative;
-  top: 1px;
-  left: 5px;
-  opacity: 0.8;
-}
-
-.kwic_sidebar {
-  background-color: #fff;
-  border-style: solid;
-  border-width: 1px;
-  border-color: #999;
-}
-
-#sidebar {
-  div {
-    margin-bottom: 10px;
-  }
-  p {
-    margin: 0 0 0 0;
-  }
-  ul {
-    margin-top: 0;
-    margin-bottom: 2px;
-    padding-left: 20px;
-    list-style: none;
-    line-height: 16px;
-  }
-
-  #selected_sentence > h4:last-child {
-    display : none;
-  }
-
-  #selected_word > h4:last-child {
-    display : none;
-  }
+  will-change: width;
 }
 
 .video-modal-bootstrap {
