+++ package.json
@@ -18,15 +18,16 @@
         "jquery-flot": "0.8.3",
         "json3": "3.3.2",
         "jstorage": "andris9/jStorage#0.4.12",
+        "leaflet": "^1.6.0",
+        "leaflet-providers": "^1.9.1",
+        "leaflet.markercluster": "^1.4.1",
         "lodash": "^4.17.19",
         "moment": "2.20.1",
         "raphael": "2.2.7",
+        "raw-loader": "^4.0.2",
         "rickshaw": "1.6.4",
         "slickgrid": "spraakbanken/SlickGrid#master",
-        "tailwindcss": "^1.6.2",
-        "leaflet": "^1.6.0",
-        "leaflet-providers": "^1.9.1",
-        "leaflet.markercluster": "^1.4.1"
+        "tailwindcss": "^1.6.2"
     },
     "devDependencies": {
         "@babel/core": "^7.2.0",
@@ -42,13 +43,13 @@
         "copy-webpack-plugin": "^6.0.3",
         "css-hot-loader": "^1.3.9",
         "css-loader": "^0.28.9",
-        "extract-loader": "^1.0.2",
         "eslint": "^7.8.1",
         "eslint-config-standard": "^14.1.1",
         "eslint-plugin-import": "^2.22.0",
         "eslint-plugin-node": "^11.1.0",
         "eslint-plugin-promise": "^4.2.1",
         "eslint-plugin-standard": "^4.0.1",
+        "extract-loader": "^1.0.2",
         "file-loader": "^1.1.11",
         "html-loader": "^0.5.5",
         "html-webpack-plugin": "^4.3.0",
@@ -69,6 +70,7 @@
         "sass-loader": "^6.0.6",
         "style-loader": "^0.20.1",
         "ts-loader": "^5.3.1",
+        "tslint": "^6.1.3",
         "tslint-config-prettier": "^1.17.0",
         "typescript": "^3.3.3333",
         "wait-on": "^5.2.0",
